fix(csp): Remove build CSP violation
[This](https://github.com/webpack/webpack/blob/ed50812a6cd47fde23f142220a414c810ff38dc5/buildin/global.js) file was causing CSP violation reports

It is included [here](https://github.com/webpack/webpack/blob/master/lib/node/NodeSourcePlugin.js#L52)
