+++ package.json
@@ -39,8 +39,8 @@
     "babel-plugin-transform-class-properties": "6.24.1",
     "babel-plugin-transform-object-rest-spread": "6.23.0",
     "babel-plugin-transform-runtime": "6.23.0",
-    "babel-runtime": "6.23.0",
-    "bundlesize": "0.12.1",
+    "babel-runtime": "6.25.0",
+    "bundlesize": "0.13.2",
     "clean-webpack-plugin": "0.1.16",
     "codeclimate-test-reporter": "0.5.0",
     "commitizen": "2.9.6",
@@ -59,7 +59,7 @@
     "extract-text-webpack-plugin": "3.0.0",
     "flow-bin": "0.51.0",
     "flow-copy-source": "https://github.com/arahansen/flow-copy-source#29ea3c2c0c58d32182ce953a5d7fa2ab788314d4",
-    "html-loader": "0.4.5",
+    "html-loader": "0.5.0",
     "husky": "0.14.3",
     "identity-obj-proxy": "3.0.0",
     "ignore-loader": "0.1.2",
@@ -88,7 +88,7 @@
     "storybook-addon-jsx": "4.1.1",
     "style-loader": "0.18.2",
     "stylelint": "8.0.0",
-    "webpack": "3.3.0",
+    "webpack": "3.4.1",
     "webpack-bundle-analyzer": "2.8.3"
   },
   "lint-staged": {
